completed refactor onto modifier
