+++ package.json
@@ -131,7 +131,6 @@
     "main": "addon-main.js",
     "app-js": {
       "./components/pikaday-input.js": "./dist/_app_/components/pikaday-input.js",
-      "./components/pikaday-input2.js": "./dist/_app_/components/pikaday-input2.js",
       "./components/pikaday-inputless.js": "./dist/_app_/components/pikaday-inputless.js",
       "./modifiers/pikaday.js": "./dist/_app_/modifiers/pikaday.js"
     }
