+++ src/components/pikaday-input.hbs
@@ -15,6 +15,8 @@
     yearRange=this.yearRange
     i18n=this.i18n
     firstDay=this.firstDay
+    container=@container
+    bound=@bound
   }}
   {{on 'change' this.didChange}}
   type='text'
