+++ test-app/app/templates/application.hbs
@@ -1,43 +1,56 @@
-<input type="text" {{pikaday onSelect=this.itWorked defaultDate=this.someDate setDefaultDate=true}} />
+<input
+  type='text'
+  {{pikaday
+    onSelect=this.itWorked
+    defaultDate=this.someDate
+    setDefaultDate=true
+  }}
+/>
 
-<h2 id="title">Welcome to Ember.js</h2>
+<h2 id='title'>Welcome to Ember.js</h2>
 
 {{outlet}}
 
 Standalone datepicker:
-<PikadayInput 
-  @format="dddd MMMM Do YYYY, h:mm:ss a" 
-  @theme="dark-theme" 
-  @onSelection={{action "doSomethingWithSelectedValue"}} 
+<PikadayInput
+  @format='dddd MMMM Do YYYY, h:mm:ss a'
+  @theme='dark-theme'
+  @onSelection={{action 'doSomethingWithSelectedValue'}}
 />
 
-<br>
+<br />
 
-Datepicker with minDate set to today (<Input @type="checkbox" @checked={{isMinDateSet}} />) and maxDate set to today (<Input @type="checkbox" @checked={{isMaxDateSet}} />):
-<PikadayInput 
-  @minDate={{if this.isMinDateSet this.today null}} 
-  @maxDate={{if this.isMaxDateSet this.today null}} 
-  @value={{someDate}} 
-  @onSelection={{action (mut someDate)}} 
+Datepicker with minDate set to today (<Input
+  @type='checkbox'
+  @checked={{isMinDateSet}}
+/>) and maxDate set to today (<Input
+  @type='checkbox'
+  @checked={{isMaxDateSet}}
+/>):
+<PikadayInput
+  @minDate={{if this.isMinDateSet this.today null}}
+  @maxDate={{if this.isMaxDateSet this.today null}}
+  @value={{someDate}}
+  @onSelection={{action (mut someDate)}}
 />
-<br>
+<br />
 
 Datepicker with bound and set value:
-<PikadayInput 
-  @value={{this.startDate}} 
-  @onSelection={{action (mut this.startDate)}} 
-  @format="DD.MM.YYYY" 
+<PikadayInput
+  @value={{this.startDate}}
+  @onSelection={{action (mut this.startDate)}}
+  @format='DD.MM.YYYY'
 />
 {{this.startDate}}
-<br>
-<button {{action "clearStartDate"}}>Clear Date</button>
-<br>
+<br />
+<button {{action 'clearStartDate'}}>Clear Date</button>
+<br />
 
 Datepicker already disabled before it's rendered:
-<PikadayInput @disabled="disabled" />
+<PikadayInput @disabled='disabled' />
 
-<br>
+<br />
 
 Inputless datepicker:
-<PikadayInputless @onSelection={{action (mut someDate)}} />
+<PikadayInputless2 @onSelection={{action (mut someDate)}} @value={{someDate}} />
 {{someDate}}
